Adding the structure of auxiliary sources and static background E&B fields. The correct computations are added but nothing being done at the moment. Instead the moment app just allocates some zero-ed out arrays which could be added to based on user-defined inputs.
Current thought process because these flops are cheap is to not include any logic within the advance method and just assume these things are present. To save on memory, the app could create small dummy arrays that are filled with zeros (instead of grid size arrays that are filled with zeros).