Added a macro gkyl_cu_memcpy_from_symbol that catches errors while copying the kernels pointers around. Previously this was failing silently. Now the unit test will simply abort, avoiding misleading tests.
However, clearly the table-driven copies are not working, which sucks bigly. Need to investigate why this is the cse. Overall, Power9 + GPU combo is a bit misleading as there are "ghost copies" happening under us on this platform which means what we may end up implementing here is not portable across machines that do not support this managed memory magic.