Adding the local 10 moment closure to the pressure tensor rotation via an integrating factor approach which allows for arbitrarily large damping coefficient. This approach is identical to what was done in G2, but done via folding the integrating factor into the pressure tensor rotation update directly instead of a loop over source options.
Since this update is (basically) trivial and we can use the ternary operator to define the closure parameter independent of model, I have coded up the local closure as if we will always have the local closure turned on. This need not be the case of course, but if the local closure parameter is not specified in the input file, the value will default to zero and a few unnecessary multiplications will be done on already loaded in data (a cost I believe is worth it to avoid some of the logical complexity of sources in G2).

Other changes include the addition of the wall boundary condition for 10 moment which requires zero tangent boundary condition on the off diagonal components of the pressure tensor, two new 10 moment regression tests (GEM and Burch event), and a small decrease in the end time of the 5 moment Burch regression test for ease of comparison to 10 moment.