Proposed deprecation of the penalty fluxes for collisionless Vlasov-Maxwell.
In G2, we have a Lua side option to choose between the default penalty fluxes or upwind (or recovery, though that is very experimental).

The reason to deprecate penalty fluxes is as follows:
1. The upwind fluxes are far, far more stable than the penalty fluxes. There are simulations which do not run at any feasible resolution with penalty fluxes that run stably at reasonably coarse resolutions with upwind fluxes.
2. Eliminates the need for the maxs_by_cell data structure in CPU and GPU implementation, and thus eliminates a reduction operation from the time-stepping. The only reduction which needs to be performed now is the scalar reduction to obtain the stable time-step.

The documented down sides:
1. Upwind fluxes are more expensive by at least a factor of 2, perhaps slightly more in higher dimensions (this is obviously not a reasonable downside when there are lots of simulations which flat. out. do. not. work with penalty fluxes).
2. The upwind fluxes kernels are prohibitively large for p>1 in higher dimensionality.
3. We currently do not understand whether there is an optimal set of nodes to evaluate the upwind direction at for the Serendipity basis for p >= 2. For tensor bases (Serendipity p=1, or Tensor p>1) there is a natural choice of the complete Gauss-Legendre basis and this choice seems to work remarkably well (see above note about how much more stable upwind fluxes are), but the choice of nodes requires experimentation.

Because the upwind flux kernels are prohibitively large for p>1 in 5D and 6D and our current lack of understanding of node location choices for Serendipity p>=2, THIS BRANCH ELIMINATES THE P=2 SERENDIPITY KERNELS.

These kernels can be added back in (take a few hours to generate with Maxima).

This commit breaks the ctest_hyper_dg because the Vlasov-Maxwell kernels are different now with the upwind fluxes.