New way of computing the upwind direction in Vlasov kernels. Now instead of doing: fUpwindQuad = 1/2 (f_l + f_r) + 1/2 sign(alpha)*(f_l - f_r)
we just check the sign of alpha at each quadrature point and then evaluate fUpwindQuad to the appropriate side of the surface.

Will add comments with performance differences on my Mac and on Traverse on GPUs. Not a significant performance improvement, but the kernels are much smaller (by upwards of 40%) which seems like a good thing.