New organization of kernels using header functions which interpolate onto surface quadrature points (since these will eventually be useful to other equation systems). These are inline double functions so no noticeable performenace difference.
Generalized formulation of upwind fluxes regardless of basis expansion. We now always evaluate the basis expansion on a tensor product quadrature basis and then project this tensor product basis onto the original modal basis as an intermediate step before the computation of Ghat.

For tensor bases (including p=1 Serendipity) this transformation is "trivial" in that there is a direct mapping between the two basis sets. This mapping is exactly what has been previously employed for the upwind fluxes. But for p>1 Serendipity we compute an inner product of the nodal basis with the modal surface basis functions. This slightly increases the cost of p>1 Serendipity upwind fluxes but is mathematically more sound, as there will be no aliasing in the final computation of Ghat, only the standard loss of accuracy going from a tensor basis to a Serendipity basis.