New Vlasov-Poisson kernels with corresponding modifications to the header file. These Vlasov-Poisson kernels come in two flavors: vlasov_poisson (which just takes fac_phi, the potential scaled by the corresponding factor such as q/m for plasmas or G*m for self-gravitating systems) and vlasov_poisson_extem (which takes both fac_phi and vecA, q/m*A the vector potential) for evolving systems with external fields in the electrostatic limit.
Names subject to some debate, but the basic idea is for the signatures of these two sets of kernels to be the same and for vlasov_poisson to take a NULL pointer for vecA to try to streamline the logic. Skeleton equation object added. This equation object needs the logic for switching between the two sets of kernels.

Compilation on my 4 core 2013 Macbook pro is now ~1 minute.