Proposal for the organization of the non-ideal terms (such as Braginskii and gradient-based closure for 10 moment) through the app.
The basic idea is that it is easiest to initialize the necessary equation objects in app/moments.c. Thus extend the moment_coupling struct in moments.c to have more equation objects which can be initialized depending on input parameters. We should then be able to call the corresponding advance methods in the moment_coupling_update function and pass the resulting RHS array to gkyl_moment_em_coupling_advance, where it can be accumulated in time with everything else.

NOTE: this is an idea and this does not compile. I am hoping to get feedback on this tomorrow before heavy debugging.