Adding LHDI tests and a slight rearrangement of the burch event test. Note that the burch event test runs much longer now, to around 50 inverse ion cyclotron times just like in the TenBarge et al. JGR paper.
Closure parameters are chosen to be identical between the local and gradient based closure runs for ease of comparison. NOTE: we use the same k0 value for both the local and gradient-based closure runs, but in the local closure k0 is taken as is, but in the gradient-based closure, we compute alpha = 1/k0.