Fixing uninitialized variable in computation of wave geometry. Surprizing it was actually working. Valgrind caught the error, though. Must make sure all code is valgrind clean!
Fixed MHD Orsag-Tang test to properly write out more frames (please
use time-triggers for this and do not do it by "hand"). Also reduced
resolution so it runs faster. This test suffers from positivity
issues. However, this needs to be handled at the level of the
wave_prop and moment app itself.