Changes to the LBO ctest to match what the LBO equation object should do: only update in the velocity space directions. The function calls are empty for the CDIM directions so we have to initialize things this way.
Also error in how the collision frequency was initialized at the app level. We technically need to project the collision frequency onto basis functions now, even when it is just a constant value, to avoid some of the more convoluted logic of varNu vs. constNu in G2.

Currently this projection is kind of messy in what is required in the input files, so this should have some abstraction away from the user. But we still have to figure out what the issue is with the G0 LBO before we worry about aesthetically pleasing input files.