Euler multiblock regression now run in parallel with job_pool. Here, each block is run in a different thread. For a 3 block problem this is giving a 2.5x speedup (block size is modest 128x128). The code structure is in place and now a decision needs to be made if to use job_pool throughout or not.
Basically, the advantage of a job_pool approach is that a pool can be
attached to each species, block etc, hence maximizing
parallelism (mixing threads, MPI and GPUs also). However, the code is
a bit more tedious as each task one wants to run in parallel needs to
be added to the job_pool. The multiblock regression went from 550 LOC
to 650 LOC. So 100 extra lines (20% increase in LOC to allow each
block to run on its own thread).