THE LBO WORKS.
Bug fix in the app for the LBO. The calls on lines 152 and 153 of vm_species.c are placeholders for now, but this does actually set a "nuVthSqSum" and "nuUSum" for use by the LBO. These need to be replaced by the primitive moment calculations and bin_ops to multiply nu, the collision frequency, by the resulting primitive moments, but things are at least being set up correctly.

Two new regression tests in 1x1v for the LBO. One is simple top hat relaxation (no drift) and the other is an initial Maxwellian. The latter test is currently "working" but needs the primitive moment computation for the final state to be the actual equilibrium for our discrete LBO.