First pass through adding the applied acceleration, currents, and external EM functionality. DOES NOT WORK YET.
I think this is the right approach that we can specify a function in the input file, and then the species and field structs have new projection updaters which will be called if it detects the presence of these functions. Currently there is no logic for NOT calling the advance methods in the moment_coupling_update method. I would prefer the proj_advance calls just do nothing if the projection object is empty, but that does not seem to be a permissible way to do things.

Seeking input.