Batched linear solves now work on GPU using CUBLAS!
There are some strange things which I do not understand fully yet,
though. The CuBLAS interface is nasty and there are some funky memory
errors that valgrind reports. However, it is possible that valgrind is
merely confused with the CUDA symbols in the exectuables.

Further work is needed to clean up the code. In particular, perhaps
all the allocations should be done once. However, this needs the
div_op design to be completed now so we can decide a proper approach
to allocating memory.