Refactoring the LBO code so that the drag and diffusion eqns are now inside the LBO updater object and not in the app species object. This is cleaner and actually the correct thing to do. Updaters should manage the eqn objects. Hyper_dg can't do this as the eqn needs to be constructed outside as it does not know what eqn we are solving. OTOH, LBO knows it needs a drag and diff eqn and so should be self-contained.
For some strange reason the LBO unit tests (on host) were failing with
nvcc. There is a somewhat larger floating point error (1e-11 v/s
1e-12). Adjusted now.

There are some unresolved issues in my mind about naming (minor) and
how zero_flux_flags are set in the dg_lbo_updater new method. AFAICT
this code is not correct and does nothing as the zero_flux is applied
in the conf directions, while the update loop is really not over conf
direction for LBO at all! I think this is a bug and needs to be fixed.

Other issues with the code are in Issue 22 on
Github. See: https://github.com/ammarhakim/gkylzero/issues/22#issue-1091863341