Progress on debugging Braginskii implementation. It seems that there is some sensitivity to floating point precision in the Braginskii implementation, as runs that were failing before with real units (SI) now run fine with a reasonable normalization (and the same ratios as the real units run).
The key to a normalized Braginskii run is that the coll_fac input parameter must be the plasma parameter (some number >> 1 determining how weakly coupled the plasma is). This scales the collision time such that the ratio of the collision time to the inverse plasma frequency is the plasma parameter, as it should be (roughly).

The Braginskii implementation still seems to require *very* small time-steps, factors of 1000-10000 smaller than the hyperbolic step for fairly coarse resolutions. More testing required on this front.