SuperLU was not being built with optimizations turned on. Also add -g flag to SLU build. Now valgrind pinpoints the proper place where the memory leak occurs. Leak seems to be inside SLU itself, but it is possible we are not calling the proper release method.
SLU docs suck bigly. Only way to find out what is happeing is to read
the source code! Weirdly SLU does not document the API in the headers,
as normal humans would, but only in the C file. So this makes it even
harder as one now needs to look at the source instead of the installed
headers. Bad design!