Adding skip_dirs to moments app to avoid updating certain directions. This is to allow axisymmetric problems on wedge domains.
However, it seems there is some problem in either my understanding or
the code as axisymmetric problem with 1 cell in theta in a narrow
wedge does not work, and in fact it looks exactly like a slab problem!
Need to think about this more, but committing for now.

Updated Moments app wrapper to cleanup memory leak.