IT LIVES! With the addition of the relativistically correct moment calculation routine, we are now getting at least more reasonable looking evolution of the relativistic two-stream instability.
Need to discuss various infrastructure here to see if there might be a better way to handle passing p/gamma to the moment calculation routine. Right now we create the moment type, set the auxiliary field, create the solver, then release the moment type (we do not hang on to the moment type). I believe this works fine but most testing is needed.