Adding capability to Vlasov solver to use a fluid advection species for an additional force in the update. This formulation allows one to include a fluid analogue of the mirror force which goes like p_perp/B*dB/dz, where p_perp/B is the fluid variable being evolved (really J*p_perp, where J=1/B is the Jacobian from transforming the equation to be along a field line).
Right now just uses the full dE/dt = -J field equation. If the system supports a steady state, we should be able to just run the simulation long enough for the plasma oscillations to damp away. If this works, can compare this solution to a solution with E coming from an Ohm's Law like Kulsrud's E_parallel equation.