Now that BCs are working on GPUs, adding another important BC to the options: absorbing BCs. 1x1v, no collisions, sheath tested added. Seems to be working (could use help from the sheath experts).
Some clean up of BCs done as well. Most of the ctxs for BC functions are identical and all the private headers include dg_eqn.h so the context functions have been unified. Also the BC function release is agnostic to the BC type so now just a single release function (in fact the only thing that is different between different BCs is the specific function; one wonders if actually these functions should be split in some way to reduce code duplication).