New entry in the dg_eqn struct "gen_surf_term" which performs a surface integral when the number of input arrays is not simply one's 1D neighbors. New equation object dg_gen_diffusion (generic diffusion) which takes as input a diffusion tensor (with off-diagonal entries) and sets the gen_surf_term function pointer to recently committed kernels by Petr for 2D anisotropic diffusion.
Restructuring of hyper_dg to have the gen_stencil methods in the gkyl_hyper_dg.h header. These objects were manipulating an identical gkyl_hyper_dg struct and thus there is no need for a new object. Now this is hooked into the fluid app with a semi-trivial change: check if the diffusion tensor is anisotropic and set the equation object and solver based on whether this bool is true.

If I have done everything correct.. this should just work