Adding FPO drag and diffusion equation objects. Since Vlasov FPO is only defined for 3V, simplify the kernel selection procedure, especially for diffusion which will require kernels for anisotropic diffusion (xx, xy, xz, yx, yy,...).
Since all the collision objects were ultimately manipulating similar structs for the dg_updater_* methods, consolidate to a similar struct. This struct still needs to be forward declared in the header of the specific file (like gkyl_dg_updater_lbo_vlasov) but now we don't need a specific struct for each different collision object that consistents of two dg_eqn structs (one for drag and one for diffusion) and two hyper_dg structs.