Clean up of mom_calc_bcorr to go back to just looping over velocity space edges, since ghost_surf_calc handles the general case of fluxes at boundaries.
Equally important: the derived constructor classes had memory leaks because they were creating types (mom_type or prim_lbo_type) and then creating the calculators which call acquire methods on these types.

As such, since we do not need the specific types subsequently after the calculator construction has acquired a pointer to these types, we call the release methods on these specific types. This decreases the reference count from 2 -> 1 (it goes from 1 -> 2 when the calculator calls the acquire) so that when the release method is called on the calculator at the end of the simulation, the reference count goes from 1 -> 0 and the specific mom_type or prim_lbo_type is free-ed.