Changing the way max speed is computed to be agnostic to rotations.
Commiting wave_prop, even though positivity does not work
fully. However, I think I understand what is happening. Basically,
testing for negative values after a first-order update is not
enough. Imagine we do a first-order update. Say it is positive. So our
test for inv domains passed. But then the second order corrections are
applied. But now the solution can be negative!

So the solution really is to update the 1D slice completely and then
check and mark edges.

This is somewhat hairy to do and needs refactor of the code. Hence, I
am commiting this intermediate state of the updater so as not to lose
any work.