gkyl_moment_species: remove the const specifier of equation
gkyl_moment_species is used in the user app and takes an equation
pointer provided by the user. It is then used in moment.c to construct
the moment_species object of the app.

By removing the const specifier, the app may change parameters of the
equation during the simulation. Presently, this is used in to update the
ch parameter of the GLM div(B) cleaning.