Significant bug fixes for the PKPM model, as well as a sizable reformulation. The bug fixes were entirely in the kinetic equation and were as follows: 1. The force terms which appear have spatial derivatives (1/rho * b . div(P) - v_parallel bb : grad(u)) which previously *did not* include the appropriate 2/dx factor to transform from computational to physical coordinates. 2. The acceleration in velocity space had the wrong derivative in the volume term. It should have been dw/dv_parallel * acceleration * f (and instead it was doing dw/dx).
By themselves, these two bug fixes did not fix all the issues we were seeing in PKPM, so a reformulation was considered.
1. Because phase space incompressibility of the kinetic equation *relies* on a cancellation of two terms (spatial advection div(u*f) and velocity space advection d/dv_parallel ( -v_parallel bb : grad(u) f), we have made u continuous using the fem_parproj routine, at least for the 1D tests being done so far.
2. With continuous u, the surface terms in both the kinetic equation and the fluid equations can be simplified significantly to use true upwinding instead of a Lax flux.
3. *Instead* of evolving a *total* energy equation, now evolve an equation for p_perp. This change reduces the number of terms which couple the fluid and kinetic equations and eliminates the need to couple the heat flux from the kinetic equation to the fluid equations.
4. In fact, with both continuous u and a p_perp equation, we no longer need the mom_pkpm_surf_calc updater. We don't need a heat flux at the surfaces, and we can use the fact that u doesn't depend on velocity space to just evaluate rho (the mass density) at the interface based on the sign of u.

Now, all three tests: sod shock, travel pulse, and square relaxation, behave reasonable with the following caveat. There is an emergent grid mode which can be most easily seen at high resolutions. Before, the grid modes which developed would kill the simulation. Now they seem tamer and much smaller amplitude (the travel pulse problem at modest resolutions does not have a visible grid mode; you would need to fourier transform or run at coarser resolution). The most striking place this happens is if you run the square_relax problem at higher configuration space resolution (say Nx = 32 cells). A mode develops when there are *zero* initial gradients despite seemingly correct relaxation of the solution to a Maxwellian (and a Maxwellian in the rest frame of the plasma as there is an initial uniform background flow).

It is possible that in order to obtain phase space incompressibility, u must be C^1 continuous because of the acceleration term which goes like d/dv_parallel (v_parallel bb : grad(u)). If u was made to be C^1 continuous, the source terms for the p_perp equation in higher dimensions would be much easier, but it remains to be seen if it is feasible to force that high degree of continuity.