Further restructuring of PKPM in preparation for generalization. Now "pkpm_vars" are computed within the kinetic side of the app, which includes computing the pkpm moments (rho, p_parallel, q_parallel) and the variables required from the EM fields such as bvar, which stores the magnetic field unit vector and magnetic field unit tensor. This change eliminates a hack in initial testing of the PKPM model that initialized a magnetic field unit vector within the species object, and now allows the regression tests of the PKPM model to include a field object (initially just B = B_0 x_hat).
Further adds a dg_calc_em_vars updater which computes various derived quantities from the electromagnetic fields, such as the aforementioned bvar, as well as the E x B velocity and what I call kappa_inv_b = b_i/kappa, the magnetic field unit vector divided by the Lorentz boost factor computed for the E x B velocity.

These derived variables from the electromagnetic fields required three new basis operations: exp_sq (compute the expansion of A^2 from A where A is a DG field), inv (compute the expansion of 1/A from A where A is a DG field), and exp_sqrt (compute the expansion of sqrt(A) from A where A is a DG field). The latter exp_sqrt currently works by projecting the DG field onto quadrature points, evaluating the square root of the quadrature evaluation, and projecting back. These three operations are only committed for a subset of dimensionalities and polynomial orders we expect to use in production simulations because these kernels are cumbersome to generate in generality. For example, the expansion of 1/A involves an analytic matrix inversion and normally these sorts of operations are handled by bin_op_div. However, in the case of things like the magnetic field unit vector, it is useful to compute the magnetic field unit tensor in a completely alias free fashion (B_i B_j/|B|^2 can be computed free of aliasing errors) and then use the square root project on that result to minimize error propagation (and in so doing reduce the number of loops over the grid since this updater computes everything in one shot).

These basis operations are also useful for computing the Lorentz boost factor from a given bulk velocity V. Will continue with the hooking in of that kernel to the special relativistic moments.

Will also work on a few unit tests of these new basis operations and compare them to our dg_bin_op routines like dot_product and div.