Before, with dg_vlasov_poisson, we got the NVLINK errors in the indirection with the volume kernels. Essentially we are doing an extra, more aesthetically pleasing indirection in the private struct (structs like dg_vlasov_poisson and dg_lbo_vlasov) where we store pointers to functions "vol" or "surf" and then those functions select the correct kernels and that function is passed to the "vol_term" and "surf_term" function in the gkyl_dg_eqn struct.
Thus, following dg_vlasov_poisson, we eliminate the function indirection in *only* the volume kernel as a test to see if this eliminates the NVLINK errors for both dg_lbo_gyrokinetic and dg_lbo_vlasov.

At some point we should consider streamlining eliminating some of these indirections as new developers might come in and wonder what the heck is going in the private header that we separate the volume kernels but not the surface kernels.