Another rearrangement of PKPM. Now instead of an equation for p_perp, we solve an equation for E_perp = 1/2 rho u_perp^2 + p_perp which allows us to rewrite the "energy" equation in PKPM in conservative form.
I am likewise switching to pure Lax fluxes for the momentum and (perpendicular) energy fluxes.

I have consolidated all the "primitive" variable calculations in pkpm into a single function which calculates u, u_perp, rhou_perp, p_perp, and p_ij in a single loop over configuration space. Likewise, the "source" terms are rearranged now that we are solving for E_perp.

It would be better to implement the operator split in fluids for these equations as right now I think the source terms for E_perp mix kinetic energy and pressure, but this will work for testing.

Now on to 2D tests..