Before, our penalization was based on the penalization for the Euler equations where the max speed is simply |u| + cs, cs = sqrt(5/3 T/m)
But PKPM has a full pressure tensor, meaning the appropriate Lax flux is actually the Lax flux for the 10 moment system |u| + sqrt(3 P_ii/rho) where P_ii is the diagonal pressure tensor component in the direction of the update.

In anisotropic plasmas such as those we are trying to model with PKPM, this will locally increase and decrease diffusion in a given direction depending on the pressure in that direction. Hopefully this will be a more accurate measure of the penalization.

It's possible there is yet another increment on this estimate that leverages the fact that our system has a collision operator. Perhaps the maximum speed is simply sqrt(P_ii/rho) or something like that, as opposed to the additional sqrt(3) factor. Tbd.