A new formulation of PKPM has been discovered which solves for both the 0th and 1st Laguerre moments by default. The logic for solving for both the 0th and 1st Laguerre moments is this formulation eliminates the need for a p_perp equation to find the T_perp normalization of the Laguerres. Instead the evolution of the 1st Laguerre moment (which is actually written as T_perp/m*(F_0 - F_1)) includes all of the necessary p_perp physics to construct the gyrotropic pressure tensor, and the associated kinetic evolution of the v_perp^2 v_parallel^k for all k moments (i.e., q_perp = v_perp^2 v_parallel, r_parperp = v_perp^2 v_parallel^2, ...).
Preliminary simulations suggest this reformulation is highly advantageous. The evolution of p_perp is more robust and simulations such as the Orszag-Tang now run without the need for hyper-diffusion.

Significant clean up is needed because the implementation of PKPM is now more complex. In fact, it is likely a good idea to move all of the PKPM implementation into its own kernel organization and associated updaters (for example, don't have PKPM use dg_updater_vlasov since PKPM requires 9 auxiliary fields while most Vlasov simulations only require 1-3).

Likewise, the signs of the various terms in the implementation need to be triple-checked. This formulation has some challenging algebra to get correct with respect to the signs and one must be careful with the signs of the source terms of the equations. In fact, it is possible there is *still* a sign error in the implementation as right now the electrons cool instead of heat in the Orszag-Tang simulation.