Adding a new benchmark which will hopefully form the capstone simulation of the letter PKPM paper (simulation of heat-flux driven instability using PKPM in high beta plasma). Needs to run on GPUs.
Also caught some old code which can be deleted. Originally I was adding a boundary condition application to the LBO solve because I was using vt^2 to penalize the jumps in the momentum solve. However, we switched to a better penalization that uses P_ii/rho (where ii is the pressure tensor in the direction of update) which allows for a more appropriate penalization based on the sound speed in the direction of the update.