moment.c: init ext_em in apply_ic_field
Previously, the first time it was initialized was in
moment_coupling_update, which happens after the first data output and
the file em_field_0.gkyl has zero values.

By init the ext_em within apply_ic_field, em_field_0.gkyl is filled with
the actual external field values.

What about app_current, which is part of the moment_field and also not
set unless the first call of moment_coupling_update? It is not part of
the output, though (though, again, maybe it should be?).