PKPM Robustness discovery: a fundamental feature of the PKPM system is that integral(v_parallel F_0 dv_parallel) = 0, where F_0 is the 0th Laguerre and Fourier Harmonic. This identity follows because we are in the local flow frame so the distribution function should not have any first moment. However, for this to hold for all time, the pressure force which arises from the streaming of the particles and the pressure force which appears in the acceleration must cancel. Thus, we can only guarantee this to be true if we make sure to discretize the pressure force and streaming term in compatible ways.
I have implemented an updater which performs the streaming update and returns the first velocity moment of the streaming update, which is equivalent to div(p_parallel b_hat), exactly what we need inside the acceleration term. Initial results are promising. I see a significant reduction in the oscillations in a variety of simulations. This also explains why different types of fluxes were giving a variety of robustness results, as some fluxes were closer in consistency to enforcing this condition (even if those fluxes were not as robust for general nonlinear problems). Will perform GPU tests in higher dimensions for further testing and think about re-factoring to reduce total number of operations.