Fixing GPU version of self-consistent div(p_parallel b_hat) calculation. Problem was this is a moment calculation and thus there are race conditions. Fixed with the same infrastructure as mom_calc (which also caused me to fix a comment in mom_calc_cu that could be more clear about why we allocate the array size we allocate).
Also commenting out the div_p write method in vlasov.c. This is a very useful debugging tool but doesn't need to be done for production calculations. Will likely delete and just leave in the github records how to do this check in case we need to do another careful look at this consistency conditions.