Experimenting with a limiter for the computation of div(p_ij) and grad_i u_j which first loops over the grid and performs a 1-cell recovery to find the cell average of div(p_ij) and grad_i u_j, then within the method that computes the gradients, checks if the cell average of div(p_ij) *or* grad_i u_j switch sign going from cell to cell (suggesting a rapid variation in p_ij and u_i). The limiter takes the form of a standard flux limiter Flux = Flux_low - phi*(Flux_low - Flux_high) where Flux_low is a low order flux (the more robust averaging procedure), Flux_high is a high order flux (the recovery procedure) and phi is the flux limiter function.
Right now a version of phi that looks especially promising is a non-standard version of the minmod limiter. The minmod limiter defaults to Flux_low if the gradients switch sign, but if the gradients do not switch sign, there can be non-trivial combinations of the low and high order flux (until the limiter judges the solution to be sufficiently smooth to default to the high order flux). *Instead* we just use the high order flux if the gradients do not switch signs, thus imposing a binary limiter instead of a continuum (either the solution at the interface is averaging or the solution at the interface is recovery, no in between). Some promising early results on 1D tests but now to check if the high resolution 2D OT (which takes a couple hours on a GPU) is stable again with this flux (it's stable with averaging, unstable with recovery, but recovery looks higher quality until it goes unstable thus motivating the development of a limiter).