Significant reorg of vm_fluid_species for maintainability and flexibility. I have repurposed the dg_diffusion equation object to be a constant diffusion equation object (scalar constant nu) with the flexibility to handle both different amounts of diffusion (grad^2, grad^4, and grad^6) and different equation types (scalar, PKPM [3 components], Isothermal Euler [4 components], and Euler [5 components]). In so doing, I have eliminated the nuHyp parameter from the PKPM recovery kernels, as no diffusion on the state variable is handled by the diffusion object, with the ability to add differing amounts of diffusion based on desired diffusivity for stability.
Now if someone wants to run with a general, spatially dependent, diffusion coefficient, one always uses dg_gen_diffusion and passes a cdim*(cdim+1) size tensor to the equation object. All of that is still handled under the hood by dg_updater_diffusion.

In the process of this reorg, I noticed lots of old code and incorrect comments and have tried to delete these for sanity as we work on the wrapping in Lua.

I also discovered the way I was doing hyper-diffusion before was incorrect, with an incorrect amount of continuity specified to accurately compute grad^4 with recovery. Will move to my workstation to test what impact this has on problems which require the hyper-diffusion, like the Orszag-Tang and GEM problems