In a number of places in the program flow, we want to compute "primitive" variables, which tend to quantities produced by some non-trivial sequence of weak divisions and weak multiplications. In the past we have handled this case by explicitly calling bin_op, sometimes many times, but this introduces serious overhead to these calculations.
An important difference between this operation and typical operations required by the LBO is that every linear solve is independent. Therefore, with batched linear solves we can actually create even larger batching, increasing performance and parallelism.

This difference forms the core idea of this new type and derived types. We will define kernels which construct combined linear solves for given equation systems, with the assumption that the input quantities are things like FiveMoments (for Vlasov) or ThreeMoments (for GK), and the output is a combined primitive variables array. In fact, this approach can further take advantage of a performance improvement which exists solely for p=1: the fact that the matrices can be analytically inverted for p=1 (which has been found to be 10-25 times faster than traditional bin_op).

Tess and I will work on the kernels and then begin testing and performance.