Some of the equation objects, such as gyrokinetics, benefit from a CFL computation which is done based on *surface* values of the phase space flux. However, the evaluation of the phase space flux inside the surfaces is an extremely costly endeavor, measured at upwards of 25% of the total update for gyrokinetics in 3x2v.
Since the phase space flux has to be evaluated at surfaces anyway, we will implement a modification to hyper_dg which returns the CFL from the surface updates. This will be highly disruptive, as it will involve modification to a large percentage of our kernels (but mostly a search and replace). Basic idea implemented in hyper_dg (does not currently build)