Large restructing of PKPM. PKPM development has been sufficiently fast-paced that a number of pieces have unintuitive names and program flow as I worked to debug the physics content of PKPM.
For example, there was a function called calc_pkpm_vars_recovery which was computing the gradient of various quantities needed as part of the PKPM update, such as bb : grad(u) and div(p).

Now the function is calc_pkpm_vars_accel and it computes the pkpm_accel_vars, which includes quantities such as div(b), bb : grad(u), the pressure force = 1/rho div(p_par b) - p_perp/rho div(b), and the vperp characteristics. Thus, the program flow is hopefully more clear and there is no misnaming of functions (since while recovery is still used for div(b) to construct a continuous b_hat, arithmetic averaging is utilized for things like grad(u)). This naming convention also made it appropriate to move the pressure computation from this function back into the fluid update so that now the PKPM Euler update looks exactly like an Euler solve, which potentially accommodates a further clean up with the Euler equations supported.

Further, as part of this refactoring, the true extension to higher polynomial orders has now been added which takes into account constraints and performance considerations. Now the basis_inv method is still used for p=1, since this method has been thoroughly tested to give the same result as weak division at a fraction of the cost, but if p>1, a weak division object is constructed which utilizes the fact that in many cases we want to do a bunch of divisions at once (bb is a 6 component tensor, there are 9 variables constructed from weak division in PKPM) so that we can parallelized over the number of components we want to solve for in addition to parallelizing over the grid. More testing on GPUs is needed (and GPU support is not currently there with this refactoring).

As part of this extension to p>1, testing revealed that only a tensor basis could satisfy properties we want, such as b . b = 1 from the computation of b_hat. Thus, for right now we only add tensor bases for 2x and 3x. It is possible a different method of construction of b_hat could be explored that would work for Serendipity kernels.

We also refactored the robustness checks so that rho, p_parallel, and p_perp are also checked at positivity control points and that the means of computing b_hat from bb is as robust as possible, first checking if |B|^2 is negative at any control points, performing the weak division, and then checking if b_x b_x, b_y b_y, or b_z b_z are negative at any of the quadrature points sampled to construct b_hat.

Preliminary testing suggests many simulations are the same to 1e-10 to 1e-12. Some simulations differ by more now that rho, p_parallel, and p_perp are being checked for robustness, suggesting our current robustness checks may be too aggressive (since these simulations were stable before). Most interestingly is that the p=2 simulations now are *more* robust from the switch from the basis inverse method to weak division, with oscillations in the electrostatic shock and nonlinear Alfven wave problem disappearing.